refactor(car): 결과 출력을 OutputView로 위임
