+++ src/main/java/racingcar/model/Car.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.model;
 
 import camp.nextstep.edu.missionutils.Randoms;
 
@@ -35,18 +35,9 @@ public class Car {
         if (getRandomNumber() >= FORWARD_STANDARD) {
             point += 1;
         }
-        print();
     }
 
     private int getRandomNumber() {
         return Randoms.pickNumberInRange(MIN_NUMBER, MAX_NUMBER);
-    }
-
-    private void print() {
-        System.out.println(name + " : " + getForward());
-    }
-
-    private String getForward() {
-        return "-".repeat(point);
     }
 }
