+++ src/main/java/racingcar/view/OutputView.java
@@ -13,6 +13,7 @@ public class OutputView {
         System.out.println(GAME_RESULT);
         for (int i = 0; i < count; i++) {
             cars.forEach(Car::forward);
+            printResult(cars);
             System.out.println();
         }
     }
@@ -21,6 +22,10 @@ public class OutputView {
         List<Car> winners = Referee.getWinner(cars);
         System.out.print(WINNER);
         System.out.println(extractWinnerName(winners));
+    }
+
+    private void printResult(List<Car> cars) {
+        cars.forEach(car -> System.out.println(car.getName() + " : " + "-".repeat(car.getPoint())));
     }
 
     private String extractWinnerName(List<Car> winners) {
